backend/requirements.txt
Flask==3.1.1
flask_cors==6.0.0
openai==1.82.1
PyJWT==2.10.1
pymongo==4.13.0
pypdf==5.6.0
python-dotenv==1.1.0
regex==2024.11.6
Requests==2.32.3
sentence_transformers==4.1.0
tiktoken==0.9.0
tqdm==4.67.1
gunicorn==21.2.0
